Move HTMLTextInput component to @wordpress/edit-post components
