+++ packages/edit-post/src/components/layout/index.native.js
@@ -16,17 +16,14 @@ import {
 	FloatingToolbar,
 } from '@wordpress/block-editor';
 import { compose, withPreferredColorScheme } from '@wordpress/compose';
-import {
-	HTMLTextInput,
-	KeyboardAvoidingView,
-	NoticeList,
-} from '@wordpress/components';
+import { KeyboardAvoidingView, NoticeList } from '@wordpress/components';
 import { AutosaveMonitor } from '@wordpress/editor';
 import { sendNativeEditorDidLayout } from '@wordpress/react-native-bridge';
 
 /**
  * Internal dependencies
  */
+import TextEditor from '../text-editor';
 import styles from './style.scss';
 import headerToolbarStyles from '../header/header-toolbar/style.scss';
 import Header from '../header';
@@ -84,7 +81,7 @@ class Layout extends Component {
 	}
 
 	renderHTML() {
-		return <HTMLTextInput parentHeight={ this.state.rootViewHeight } />;
+		return <TextEditor parentHeight={ this.state.rootViewHeight } />;
 	}
 
 	renderVisual() {
