+++ packages/edit-post/src/components/text-editor/native-container/style.native.scss
@@ -1,0 +1,12 @@
+
+.keyboardAvoidingView {
+	position: absolute;
+	top: 0;
+	right: 0;
+	left: 0;
+	bottom: 0;
+}
+
+.scrollView {
+	flex: 1;
+}
