+++ packages/edit-post/src/components/text-editor/style-common.native.scss
@@ -2,18 +2,6 @@ $padding: 8;
 $htmlFont: $default-monospace-font;
 $textColorDark: $white;
 
-.keyboardAvoidingView {
-	position: absolute;
-	top: 0;
-	right: 0;
-	left: 0;
-	bottom: 0;
-}
-
-.container {
-	flex: 1;
-}
-
 .placeholder {
 	color: $gray;
 }
