+++ packages/edit-post/src/components/text-editor/style.android.scss
@@ -15,7 +15,3 @@
 	padding-top: $padding;
 	padding-bottom: $padding;
 }
-
-.scrollView {
-	flex: 1;
-}
